+++ gila-multi/src/unipg/gila/multi/io/JavaSolarMergerTestOutputFormat.java
@@ -18,19 +18,16 @@
  */
 package unipg.gila.multi.io;
 
-import org.apache.giraph.io.formats.TextVertexOutputFormat;
-
 import java.io.IOException;
 import java.util.Iterator;
 
 import org.apache.giraph.edge.Edge;
 import org.apache.giraph.graph.Vertex;
-import org.apache.hadoop.io.FloatWritable;
+import org.apache.giraph.io.formats.TextVertexOutputFormat;
 import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.TaskAttemptContext;
 import org.apache.log4j.Logger;
-import org.mortbay.log.Log;
 
 import unipg.gila.common.coordinatewritables.AstralBodyCoordinateWritable;
 import unipg.gila.common.multi.LayeredPartitionedLongWritable;
@@ -54,9 +51,7 @@ public class JavaSolarMergerTestOutputFormat extends TextVertexOutputFormat<Laye
 					throws IOException {
 				LayeredPartitionedLongWritable id = vertex.getId();
 				AstralBodyCoordinateWritable value = vertex.getValue();
-				float[] cohords = vertex.getValue().getCoordinates();
-//				if(vertex.getValue().isMoon())
-//					log.info("Moon " + vertex.getValue().getSun() + " of layer " + vertex.getId().getLayer() + " has " + vertex.getValue().getProxies().size() + " proxies");
+				double[] cohords = vertex.getValue().getCoordinates();
 				return new Text("[" + id.getId() + "," + id.getLayer() + "," + cohords[0] + "," + cohords[1] + "," + SolarMerger.AstralBody.toString(SolarMerger.AstralBody.buildBody(value.getDistanceFromSun())) + "," + ((LayeredPartitionedLongWritable)value.getSun()).getId() + ",[" + edgeBundler(vertex.getEdges()) + "]]");
 			}
 		}
