+++ gila-multi/src/unipg/gila/multi/io/SolarMergerTestWithSpanningTreeOutputFormat.java
@@ -18,25 +18,20 @@
  */
 package unipg.gila.multi.io;
 
-import org.apache.giraph.io.formats.TextVertexOutputFormat;
-
 import java.io.IOException;
 import java.util.Iterator;
 
 import org.apache.giraph.edge.Edge;
 import org.apache.giraph.graph.Vertex;
-import org.apache.hadoop.io.FloatWritable;
-import org.apache.hadoop.io.IntWritable;
+import org.apache.giraph.io.formats.TextVertexOutputFormat;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.TaskAttemptContext;
 import org.apache.log4j.Logger;
-import org.mortbay.log.Log;
 
 import unipg.gila.common.coordinatewritables.AstralBodyCoordinateWritable;
 import unipg.gila.common.datastructures.SpTreeEdgeValue;
 import unipg.gila.common.multi.LayeredPartitionedLongWritable;
 import unipg.gila.multi.coarseners.SolarMerger;
-import unipg.gila.multi.io.SolarMergerTestOutputFormat.JavaSolarMergerTestVertexWriter;
 
 public class SolarMergerTestWithSpanningTreeOutputFormat extends TextVertexOutputFormat<LayeredPartitionedLongWritable, AstralBodyCoordinateWritable, SpTreeEdgeValue> {
 
@@ -56,7 +51,7 @@ public class SolarMergerTestWithSpanningTreeOutputFormat extends TextVertexOutpu
           throws IOException {
       LayeredPartitionedLongWritable id = vertex.getId();
       AstralBodyCoordinateWritable value = vertex.getValue();
-      float[] cohords = vertex.getValue().getCoordinates();
+      double[] cohords = vertex.getValue().getCoordinates();
       //        if(vertex.getValue().isMoon())
       //          log.info("Moon " + vertex.getValue().getSun() + " of layer " + vertex.getId().getLayer() + " has " + vertex.getValue().getProxies().size() + " proxies");
       return new Text("[" + id.getId() + "," + id.getLayer() + "," + cohords[0] + "," + cohords[1] + "," + SolarMerger.AstralBody.toString(SolarMerger.AstralBody.buildBody(value.getDistanceFromSun())) + "," + ((LayeredPartitionedLongWritable)value.getSun()).getId() + ",[" + edgeBundlerWithSpanningTree(vertex.getEdges(), false) + "]," + "[" + edgeBundlerWithSpanningTree(vertex.getEdges(), true) + "]"  + "]");
